@@ -55,7 +55,7 @@
         <dependency>
             <groupId>com.github.arynxd</groupId>
             <artifactId>button-pagination</artifactId>
-            <version>bc4174864e</version>
+            <version>master-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.jsoup</groupId>
@@ -77,13 +77,6 @@
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
             <version>1.3.0-alpha5</version>
-        </dependency>
-        <dependency>
-            <groupId>com.jagrosh</groupId>
-            <artifactId>jda-utilities</artifactId>
-            <version>3.0.5</version>
-            <scope>compile</scope>
-            <type>pom</type>
         </dependency>
         <dependency>
             <groupId>commons-io</groupId>
@@ -125,8 +118,8 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.8.1</version>
                     <configuration>
-                        <source>15</source>
-                        <target>15</target>
+                        <source>16</source>
+                        <target>16</target>
                         <compilerArgs>--enable-preview</compilerArgs>
                     </configuration>
                 </plugin>
