@@ -5,13 +5,9 @@ package schoolbot.commands.admin;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
-import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageChannel;
-import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.message.react.MessageReactionAddEvent;
 import org.jetbrains.annotations.NotNull;
-import schoolbot.Constants;
 import schoolbot.objects.command.Command;
 import schoolbot.objects.command.CommandEvent;
 import schoolbot.util.Checks;
@@ -47,10 +43,11 @@ public class Clear extends Command
             if (args.isEmpty())
             {
 
-                  MessageChannel channel = event.getChannel();
-                  Message message = event.getMessage();
-                  Member member = event.getMember();
-                  User author = event.getUser();
+                  var channel = event.getChannel();
+                  var message = event.getMessage();
+                  var member = event.getMember();
+                  var author = event.getUser();
+                  var selfUser = event.getJDA().getSelfUser();
 
                   message.delete().queue();
 
@@ -73,41 +70,33 @@ public class Clear extends Command
                         prompt.addReaction("\u2705").queue(); // Checkmark.
                         prompt.addReaction("\u274E").queue(); // X-Mark.
 
-                        event.getSchoolbot().getEventWaiter().waitForEvent(MessageReactionAddEvent.class, reactionEvent -> reactionEvent.getMessageIdLong() == prompt.getIdLong()
-                                                                                                                           && Objects.equals(reactionEvent.getUser(), author) || Objects.equals(reactionEvent.getUser().getId(), Constants.GENIUS_OWNER_ID), reactionEvent ->
-                        {
+                        event.getSchoolbot().getEventWaiter().waitForEvent(MessageReactionAddEvent.class,
+                                reactionEvent -> reactionEvent.getMessageIdLong() == prompt.getIdLong()
+                                                 && Objects.equals(reactionEvent.getUser(), author) || Objects.equals(reactionEvent.getUser().getIdLong(), selfUser.getIdLong()),
 
-                              switch (reactionEvent.getReaction().getReactionEmote().getName())
-                              {
-                                    case "\u2705" -> {
-                                          prompt.getChannel().getIterableHistory()
-                                                  .takeAsync(100)
-                                                  .thenApplyAsync(channelMessages ->
-                                                  {
-                                                        List<Message> deletableMessages = channelMessages.stream()
-                                                                .filter(messages -> messages.getTimeCreated().isBefore(
-                                                                        OffsetDateTime.now().plus(2, ChronoUnit.WEEKS)
-                                                                )).collect(Collectors.toList());
+                                reactionEvent ->
+                                {
+                                      switch (reactionEvent.getReaction().getReactionEmote().getName())
+                                      {
+                                            case "\u2705" -> prompt.getChannel().getIterableHistory()
+                                                    .takeAsync(100)
+                                                    .thenApplyAsync(channelMessages ->
+                                                    {
+                                                          List<Message> deletableMessages = channelMessages.stream()
+                                                                  .filter(messages -> messages.getTimeCreated().isBefore(
+                                                                          OffsetDateTime.now().plus(2, ChronoUnit.WEEKS)
+                                                                  )).collect(Collectors.toList());
 
-                                                        channel.purgeMessages(deletableMessages);
+                                                          channel.purgeMessages(deletableMessages);
 
-                                                        return deletableMessages.size();
-                                                  }).whenCompleteAsync((messagesTotal, throwable) -> channel.sendMessage(
-                                                  "Successfully purged `" + messagesTotal + "` messages."
-                                          ).queue(botMessage -> botMessage.delete().queueAfter(5, TimeUnit.SECONDS)));
-                                          break;
-                                    }
-                                    case "\u274E" -> {
-                                          channel.sendMessage("Operation was successfully cancelled.").queue();
-                                          break;
-                                    }
-                                    default -> {
-                                          channel.sendMessage("You did not select one of the available options.").queue();
-
-                                          break;
-                                    }
-                              }
-                        }, 15, TimeUnit.SECONDS, () ->
+                                                          return deletableMessages.size();
+                                                    }).whenCompleteAsync((messagesTotal, throwable) -> channel.sendMessage(
+                                                            "Successfully purged `" + messagesTotal + "` messages."
+                                                    ).queue(botMessage -> botMessage.delete().queueAfter(5, TimeUnit.SECONDS)));
+                                            case "\u274E" -> channel.sendMessage("Operation was successfully cancelled.").queue();
+                                            default -> channel.sendMessage("You did not select one of the available options.").queue();
+                                      }
+                                }, 15, TimeUnit.SECONDS, () ->
                         {
                               prompt.delete().queue();
 
