@@ -5,11 +5,13 @@ import java.awt.*;
 public class Constants
 {
       public final static String DEFAULT_PREFIX = ".";
-      public final static Color DEFAULT_EMBED_COLOR = Color.GRAY;
+      public final static Color DEFAULT_EMBED_COLOR = Color.BLACK;
       public final static long GENIUS_OWNER_ID = 105141507996061696L;
       public final static String VERSION = "1.0.0";
       public final static int MAX_GUILD_ROLE_COUNT = 250;
       public final static int MAX_GUILD_TEXTCHANNEL_COUNT = 500;
+      public final static int MAX_EMBED_TITLE = 256;
+      public final static int MAX_FIELD_VALUE = 1024;
 
 
 }
