@@ -9,15 +9,17 @@ import org.jsoup.Jsoup;
 import schoolbot.objects.command.Command;
 import schoolbot.objects.command.CommandEvent;
 import schoolbot.objects.command.CommandFlag;
+import schoolbot.objects.misc.StateMachine;
+import schoolbot.objects.misc.StateMachineValues;
 import schoolbot.objects.school.Classroom;
 import schoolbot.objects.school.School;
 import schoolbot.util.Checks;
 import schoolbot.util.Embed;
+import schoolbot.util.Processor;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 
 public class ClassroomAdd extends Command
@@ -49,189 +51,42 @@ public class ClassroomAdd extends Command
 
       }
 
-      public static class ClassAddStateMachine extends ListenerAdapter
+      /**
+       * Returns true if PeopleSoft is up and functioning if not it return false
+       * The values parameter is all of the state machines possible values
+       *
+       * @param values All of the state machines possible values
+       * @return False is PeopleSoft is up and functioning. Otherwise fails
+       */
+      private static boolean isDown(@NotNull StateMachineValues values)
       {
-            private final long channelID, authorID;
-            private int state = 1;
-            private final CommandEvent commandEvent;
-            private String CLASS_SEARCH_URL = "https://psmobile.pitt.edu/app/catalog/classsection/UPITT/";
-            private Classroom schoolClass;
+            var jda = values.getJda();
+            var event = values.getEvent();
+            var machine = values.getMachine();
 
-            public ClassAddStateMachine(CommandEvent event)
+            Connection.Response response;
+            try
             {
-                  this.channelID = event.getChannel().getIdLong();
-                  this.authorID = event.getUser().getIdLong();
-                  this.commandEvent = event;
+                  response = Jsoup.connect("https://psmobile.pitt.edu/app/catalog/classSearch")
+                          .followRedirects(true)
+                          .execute();
+
+            }
+            catch (Exception e)
+            {
+                  Embed.error(event, "Error while attempting to connect to PeopleSoft");
+                  jda.removeEventListener(machine);
+                  return true;
             }
 
-            @Override
-            public void onGuildMessageReceived(@NotNull GuildMessageReceivedEvent event)
+
+            if (response.url().toString().equals("https://prd.ps.pitt.edu/Maintenance.html"))
             {
-                  if (event.getAuthor().isBot()) return;
-                  if (event.getAuthor().getIdLong() != authorID) return;
-                  if (event.getChannel().getIdLong() != channelID) return;
-
-
-                  String message = event.getMessage().getContentRaw();
-                  List<School> cachedSchools = commandEvent.getGuildSchools();
-                  List<School> pittSchools = commandEvent.getGuildSchools()
-                          .stream()
-                          .filter(School::isPittSchool)
-                          .collect(Collectors.toList());
-
-                  var jda = event.getJDA();
-                  var guild = event.getGuild();
-                  var channel = event.getChannel();
-
-
-                  if (message.equalsIgnoreCase("stop"))
-                  {
-                        channel.sendMessage("Okay aborting..").queue();
-                        jda.removeEventListener(this);
-                        return;
-                  }
-
-                  // TODO: Put in support for other schools other than pitt classes
-
-
-                  switch (state)
-                  {
-                        case 1 -> {
-                              schoolClass = new Classroom();
-                              schoolClass.setGuildID(guild.getIdLong());
-
-                              if (message.equalsIgnoreCase("Yes") || message.equalsIgnoreCase("y"))
-                              {
-                                    Connection.Response response = null;
-                                    try
-                                    {
-                                          response = Jsoup.connect("https://psmobile.pitt.edu/app/catalog/classSearch")
-                                                  .followRedirects(true)
-                                                  .execute();
-
-                                    }
-                                    catch (Exception e)
-                                    {
-                                          e.printStackTrace();
-                                          return;
-                                    }
-
-
-                                    if (response.url().toString().equals("https://prd.ps.pitt.edu/Maintenance.html"))
-                                    {
-                                          Embed.error(event, "People soft is currently down for maintenance");
-                                          jda.removeEventListener(this);
-                                          return;
-                                    }
-
-                                    if (pittSchools.isEmpty())
-                                    {
-                                          Embed.error(event, "{} has no pitt schools", event.getGuild().getName());
-                                    }
-                                    else if (pittSchools.size() == 1)
-                                    {
-                                          School school = pittSchools.get(0);
-                                          this.schoolClass.setSchool(school);
-                                          event.getChannel().sendMessage(school.getAsEmbed(commandEvent.getSchoolbot())).queue();
-                                          event.getChannel().sendMessage(school.getName() + " has automatically been selected because it is the only school available").queue();
-                                          channel.sendMessage("""
-                                                  I will now need your term. I only understand pitt term like
-                                                  ```
-                                                  Fall 2021
-                                                  Spring 2020
-                                                  Summer 2019
-                                                                                
-                                                  Format: <Season> <Year number>
-                                                  ```
-                                                  """).queue();
-                                          state = 3;
-                                    }
-                                    else
-                                    {
-                                          commandEvent.sendAsPaginatorWithPageNumbers(pittSchools);
-                                          commandEvent.sendMessage("Please pick the campus based off the page numbers :)");
-                                          state = 2;
-                                    }
-                              }
-                              else if (message.equalsIgnoreCase("no") || message.equalsIgnoreCase("nah"))
-                              {
-                                    // other school logic
-                              }
-                              else
-                              {
-                                    channel.sendMessageFormat("** %s ** is not a yes or no answer.. Try again", message).queue();
-                              }
-
-
-                        }
-                        case 2 -> {
-                              if (!Checks.isNumber(message))
-                              {
-                                    Embed.error(commandEvent, "%s is not a number", message);
-                                    return;
-                              }
-
-
-                              int number = Integer.parseInt(message);
-
-                              if (!Checks.between(number, pittSchools.size()))
-                              {
-                                    Embed.error(event, "%d is not in between 1 and %d", number, pittSchools.size());
-                                    return;
-                              }
-                              schoolClass.setSchool(pittSchools.get(number - 1));
-                              Embed.success(event, "Successfully set school to %s", schoolClass.getSchool().getName());
-                              channel.sendMessage("""
-                                      I will now need your term. I only understand pitt term like
-                                      ```
-                                      Fall 2021
-                                      Spring 2020
-                                      Summer 2019
-                                                                    
-                                      Format: <Season> <Year number>
-                                      ```
-                                      """).queue();
-                              state = 3;
-                        }
-                        case 3 -> {
-                              int term = termValidator(message);
-                              if (term == -1)
-                              {
-                                    Embed.error(event, """
-                                            Not a valid term. Aborting..
-                                            Reason for Aborting
-                                            1. **Term is either to old or too far ahead in the future**
-                                            2. **You mistyped the term**
-                                            3. **You did not input a valid season**""");
-                                    jda.removeEventListener(this);
-                                    state = 1;
-                                    break;
-                              }
-                              schoolClass.setTerm(termFixed(message));
-                              CLASS_SEARCH_URL += term + "/";
-                              channel.sendMessage("""
-                                      What is your class number
-                                      `Hint: This can normally be found on your Syllabus, PsMobile or PeopleSoft, or in the link of your class`
-                                      """).queue();
-                              state = 4;
-                        }
-                        case 4 -> {
-
-                              if (!Checks.isNumber(message))
-                              {
-                                    Embed.notANumberError(event, message);
-                                    return;
-                              }
-                              CLASS_SEARCH_URL += message;
-
-                              School school = schoolClass.getSchool();
-                              schoolClass.setURL(CLASS_SEARCH_URL);
-                              schoolClass.setNumber(Integer.parseInt(message));
-                              commandEvent.getCommandThreadPool().execute(() -> school.addPittClass(commandEvent, schoolClass));
-                              event.getJDA().removeEventListener(this);
-                        }
-                  }
+                  Embed.error(event, "People soft is currently down for maintenance");
+                  jda.removeEventListener(machine);
+                  return true;
             }
+            return false;
       }
 
 
@@ -283,5 +138,175 @@ public class ClassroomAdd extends Command
             }
 
             return String.valueOf(termCharArr) + " " + term.split("\\s+")[1];
+      }
+
+      public static class ClassAddStateMachine extends ListenerAdapter implements StateMachine
+      {
+            private int state = 1;
+            private String CLASS_SEARCH_URL = "https://psmobile.pitt.edu/app/catalog/classsection/UPITT/";
+            private final StateMachineValues values;
+
+            public ClassAddStateMachine(CommandEvent event)
+            {
+                  values = new StateMachineValues(event, this);
+            }
+
+            @Override
+            public void onGuildMessageReceived(@NotNull GuildMessageReceivedEvent event)
+            {
+
+                  long authorId = values.getAuthorId();
+                  long channelId = values.getChannelId();
+
+                  if (!Checks.eventMeetsPrerequisites(event, this, channelId, authorId))
+                  {
+                        return;
+                  }
+
+                  String message = event.getMessage().getContentRaw();
+                  var jda = values.getJda();
+                  var channel = event.getChannel();
+                  var commandEvent = values.getEvent();
+
+                  // TODO: Put in support for other schools other than pitt classes
+
+
+                  switch (state)
+                  {
+                        case 1 -> {
+                              values.setClassroom(new Classroom(event.getGuild().getIdLong()));
+
+                              if (message.equalsIgnoreCase("Yes") || message.equalsIgnoreCase("y"))
+                              {
+                                    if (isDown(values))
+                                    {
+                                          return;
+                                    }
+
+                                    var success = Processor.processGenericList(values, values.getPittClass(), School.class);
+
+                                    if (success == 1)
+                                    {
+                                          var school = values.getSchool();
+                                          channel.sendMessageFormat("""
+                                                                                                    
+                                                  ** %s ** has been selected successfully. I will now need your term. I only understand pitt term like
+                                                  ```
+                                                  Fall 2021
+                                                  Spring 2020
+                                                  Summer 2019
+                                                                                
+                                                  Format: <Season> <Year number>
+                                                  ```
+                                                  """, school.getName()).queue();
+                                          values.getClassroom().setSchool(school);
+                                          state = 3;
+                                          return;
+                                    }
+                                    else if (success == 2) state = 2;
+                                    // else case here dont forget
+
+
+                              }
+                              else if (message.equalsIgnoreCase("no") || message.equalsIgnoreCase("nah"))
+                              {
+                                    channel.sendMessage("We will start by getting the school you want to add the class to").queue();
+                                    var schools = Processor.processGenericList(values, values.getSchoolList(), School.class);
+
+                                    if (schools == 1)
+                                    {
+                                          commandEvent.sendMessage("This school has been selected because there is only one available");
+                                    }
+                                    else if (schools > 1)
+                                    {
+                                          commandEvent.sendMessage("Please select a school from the page numbers");
+                                          state = 10;
+                                    }
+                              }
+                              else
+                              {
+                                    channel.sendMessageFormat("** %s ** is not a yes or no answer.. Try again", message).queue();
+                              }
+
+
+                        }
+                        case 2 -> {
+
+                              var pittClasses = values.getPittClass();
+                              var success = Processor.validateMessage(event, pittClasses);
+
+                              if (success == null)
+                              {
+                                    return;
+                              }
+
+                              values.getClassroom().setSchool(success);
+
+
+                              Embed.success(event, "Successfully set school to %s", success.getName());
+                              channel.sendMessage("""
+                                      I will now need your term. I only understand pitt term like
+                                      ```
+                                      Fall 2021
+                                      Spring 2020
+                                      Summer 2019
+                                                                    
+                                      Format: <Season> <Year number>
+                                      ```
+                                      """).queue();
+                              state = 3;
+                        }
+                        case 3 -> {
+                              int term = termValidator(message);
+                              if (term == -1)
+                              {
+                                    Embed.error(event, """
+                                            Not a valid term. Aborting..
+                                            Reason for Aborting
+                                            1. **Term is either to old or too far ahead in the future**
+                                            2. **You mistyped the term**
+                                            3. **You did not input a valid season**""");
+                                    jda.removeEventListener(this);
+                                    state = 1;
+                                    break;
+                              }
+                              values.getClassroom().setTerm(termFixed(message));
+                              CLASS_SEARCH_URL += term + "/";
+                              channel.sendMessage("""
+                                      What is your class number. `Hint: This can normally be found on your Syllabus, PsMobile or PeopleSoft, or in the link of your class`
+                                      """).queue();
+                              state = 4;
+                        }
+                        case 4 -> {
+
+
+                              if (!Checks.isNumber(message))
+                              {
+                                    Embed.notANumberError(event, message);
+                                    return;
+                              }
+
+                              CLASS_SEARCH_URL += message;
+
+
+                              values.getClassroom().setURL(CLASS_SEARCH_URL);
+                              values.getClassroom().setNumber(Integer.parseInt(message));
+
+                              final var classroom = values.getClassroom();
+                              final var school = classroom.getSchool();
+
+
+                              commandEvent.getCommandThreadPool().execute(() -> school.addPittClass(commandEvent, classroom));
+
+                              event.getJDA().removeEventListener(this);
+                        }
+
+                  }
+
+
+                  // Custom school states
+
+            }
+
       }
 }
