@@ -3,6 +3,6 @@ package schoolbot.objects.misc;
 /**
  * Used to transfer data between class when updating classes, schools, assignments, etc
  */
-public record DatabaseDTO(Object obj, String updateColumn, Object value)
+public record DatabaseDTO(Object objectBeingUpdated, String updateColumn, Object valueBeingChanged)
 {
 }
