@@ -57,13 +57,16 @@ public class Parser
             return args;
       }
 
-      public static void classTime(Schoolbot schoolbot, String time, Classroom classroom)
+      public static boolean classTime(Schoolbot schoolbot, String time, Classroom classroom)
       {
 
-            if (time.isEmpty() || time.isBlank())
-            {
-                  return;
-            }
+            if (time.isEmpty() || time.isBlank()) return false;
+            if (!time.contains(":")) return false;
+            if (!time.contains("am") && !time.contains("pm")) return false;
+            // too lazy to do negation rn
+            if (time.contains("Mo") || time.contains("Tu") || time.contains("We") || time.contains("Th") || time.contains("Fr"))
+                  ;
+            else return false;
 
 
             Map<DayOfWeek, LocalDateTime> stuff = parseTime(classroom, time);
@@ -71,7 +74,27 @@ public class Parser
 
             LocalDate ld = classroom.getStartDate().isBefore(LocalDate.now()) ? LocalDate.now() : classroom.getStartDate();
 
+            DatabaseUtil.removeClassReminderByClass(schoolbot, classroom);
+
             // Could speed this up by only jumping to valid days
+            // TODO: Switch to while loop and only jump to days we need.
+
+            while (ld.isBefore(classroom.getEndDate()))
+            {
+                  if (ld.isAfter(LocalDate.now()))
+                  {
+                        DayOfWeek day = ld.getDayOfWeek();
+                        if (stuff.containsKey(day))
+                        {
+                              LocalTime localTime = stuff.get(day).toLocalTime();
+
+                              DatabaseUtil.addClassReminder(schoolbot, LocalDateTime.of(ld, localTime), List.of(60, 30, 10), classroom);
+
+                        }
+                  }
+                  ld = ld.plusDays(1);
+            }
+            /*
             for (; ld.isBefore(classroom.getEndDate()); ld = ld.plusDays(1))
             {
                   if (ld.isAfter(LocalDate.now()))
@@ -84,6 +107,42 @@ public class Parser
                         }
                   }
             }
+             */
+            return true;
+      }
+
+      public static boolean parseClassEditTime(String potTime, Classroom classroom)
+      {
+            Map<DayOfWeek, LocalDateTime> dayz = new HashMap<>();
+            Map<String, DayOfWeek> stringDayOfWeekMap = Map.of(
+                    "Mo", DayOfWeek.MONDAY,
+                    "Tu", DayOfWeek.TUESDAY,
+                    "We", DayOfWeek.WEDNESDAY,
+                    "Th", DayOfWeek.THURSDAY,
+                    "Fr", DayOfWeek.FRIDAY
+            );
+
+            String time = potTime.toLowerCase();
+
+            if (time.isBlank() || time.isBlank()) return false;
+            if (!time.contains(":")) return false;
+            if (!time.contains("am") && !time.contains("pm")) return false;
+            if (!potTime.contains("Mo") && !potTime.contains("Tu") && !potTime.contains("We") && !potTime.contains("Th") && !potTime.contains("Fr"))
+                  return false;
+
+
+            String[] initialSplit = potTime.split("\\s+");
+            String days = initialSplit[0];
+            String[] daysSplit = days.split("(?=\\p{Upper})");
+
+            for (String day : daysSplit)
+            {
+                  if (stringDayOfWeekMap.containsKey(day))
+                  {
+                        //dayz.put()
+                  }
+            }
+            return false;
       }
 
       private static Map<DayOfWeek, LocalDateTime> parseTime(Classroom classroom, String time)
@@ -112,8 +171,8 @@ public class Parser
             String classTime = initialSplit[1].toLowerCase();
             String[] daysSplit = days.split("(?=\\p{Upper})");
 
-            int hour = 0;
-            int minute = 0;
+            int hour;
+            int minute;
 
             String[] classTimeSplit = classTime.split(":");
 
